+++ src/tools/ToolsTBC/vmap_tools/vmap4_extractor/vmapexport.cpp
@@ -54,7 +54,7 @@ typedef struct
 
 map_id * map_ids;
 uint16 *LiqType = 0;
-uint32 map_count;
+size_t map_count;
 char output_path[128]=".";
 char input_path[1024]=".";
 bool hasInputPathParam = false;
@@ -491,7 +491,7 @@ int main(int argc, char ** argv)
             printf("FATAL ERROR: Map.dbc not found in data file.\n");
             return 1;
         }
-        map_count=dbc->getRecordCount ();
+        map_count = dbc->getRecordCount();
         map_ids=new map_id[map_count];
         for (unsigned int x=0;x<map_count;++x)
         {
